changes for pylint feedback
