+++ .pylintrc
@@ -2,7 +2,7 @@
 ignore=__pycache__,data/MNIST
 
 [MESSAGES CONTROL]
-disable=missing-docstring,invalid-name,too-few-public-methods
+disable=missing-docstring,invalid-name,too-few-public-methods,too-many-positional-arguments
 
 [FORMAT]
 max-line-length=120
@@ -23,4 +23,3 @@ max-module-lines=1000
 [REPORTS]
 output-format=colorized
 reports=no
-
