+++ labs/lab2.py
@@ -1,16 +1,14 @@
 import logging
 import os.path
 
-import torch
-
 import matplotlib.pyplot as plt
 from torch.utils.data import DataLoader
 from torch import nn
-from torch.optim import Adam
 from models.data.data_set import DataSet
 from tools.data import get_asl_data_set
 from tools.device import get_device
 from tools.helper_system import get_labs_data_saving_dir
+from tools.training import compile_model, train_and_validate_model
 
 # Constants
 IMG_SIZE = 28
@@ -99,18 +97,20 @@ def lab2():
         # Build the neural network model
         logging.info("Creating model layers and compiling the model.")
         layers = _get_model_layers()
-        model = _compile_model(device, layers)
+        model = compile_model(device, layers)
         loss_function = nn.CrossEntropyLoss()
         logging.info("Model and loss function ready.")
 
         # Train and validate the model for a number of epochs
-        logging.info("Starting training and validation for %d epochs.", EPOCHS)
-        for epoch in range(EPOCHS):
-            logging.info("Epoch %d/%d - Training phase", epoch + 1, EPOCHS)
-            _train_model(model, train_loader, device, loss_function)
-            logging.info("Epoch %d/%d - Validation phase", epoch + 1, EPOCHS)
-            _validate_model(model, valid_loader, loss_function)
-        logging.info("Training and validation complete.")
+        train_and_validate_model(
+            epochs=EPOCHS,
+            model=model,
+            train_loader=train_loader,
+            valid_loader=valid_loader,
+            device=device,
+            loss_function=loss_function,
+            is_lab1=True
+        )
 
     except (RuntimeError, ValueError, TypeError) as e:
         logging.error("Error in lab2: %s", e)
@@ -137,67 +137,6 @@ def _get_model_layers() -> list:
         IMG_SIZE, IMG_SIZE, NUM_CLASSES, layers
     )
     return layers
-
-
-def _compile_model(device, layers: list) -> nn.Module:
-    """
-    Compiles and returns a PyTorch sequential model on the specified device.
-    """
-    model = nn.Sequential(*layers)
-    model.to(device)
-    try:
-        model = torch.compile(model)
-        logging.info("Model compiled with torch.compile. Model layers: %s", layers)
-    except (RuntimeError, TypeError) as e:
-        logging.warning("torch.compile failed: %s. Using uncompiled model.", e)
-    logging.info(
-        "Model device: %s",
-        next(model.parameters()).device
-    )
-    return model
-
-
-def _train_model(model: nn.Module, loader: DataLoader, device, loss_function) -> None:
-    """
-    Trains the model for one epoch using the training data loader.
-    """
-    optimizer = Adam(model.parameters())
-    train_n = len(loader.dataset)
-    loss = 0.0
-    accuracy = 0.0
-    model.train()
-    for x, y in loader:
-        x, y = x.to(device), y.to(device)
-        output = model(x)
-        optimizer.zero_grad()
-        batch_loss = loss_function(output, y)
-        batch_loss.backward()
-        optimizer.step()
-        loss += batch_loss.item()
-        accuracy += get_batch_accuracy(output, y, train_n)
-    logging.info('Train - Loss: %.4f Accuracy: %.4f', loss, accuracy)
-
-
-def _validate_model(model: nn.Module, loader: DataLoader, loss_function) -> None:
-    """
-    Validates the model using the validation data loader.
-    """
-    loss = 0.0
-    accuracy = 0.0
-    valid_N = len(loader.dataset)
-    model.eval()
-    with torch.no_grad():
-        for x, y in loader:
-            output = model(x)
-            loss += loss_function(output, y).item()
-            accuracy += get_batch_accuracy(output, y, valid_N)
-    logging.info('Valid - Loss: %.4f Accuracy: %.4f', loss, accuracy)
-
-
-def get_batch_accuracy(output, y, N):
-    pred = output.argmax(dim=1, keepdim=True)
-    correct = pred.eq(y.view_as(pred)).sum().item()
-    return correct / N
 
 
 def _get_train_data_and_loader(x_df, y_df, device, batch_size, shuffle=True):
